Restructure modules based on protocols
This moves code into one of the following packages:

- postgres
- http
- ssl
