+++ http/src/main/java/io/crate/plugin/HttpTransportPlugin.java
@@ -21,7 +21,7 @@
 
 package io.crate.plugin;
 
-import io.crate.http.netty.CrateNettyHttpServerTransport;
+import io.crate.protocols.http.CrateNettyHttpServerTransport;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
