+++ sql/src/main/java/io/crate/protocols/postgres/PostgresWireProtocol.java
@@ -28,12 +28,11 @@ import io.crate.action.sql.SQLOperations;
 import io.crate.action.sql.SessionContext;
 import io.crate.analyze.symbol.Field;
 import io.crate.analyze.symbol.Symbols;
-import io.crate.http.netty.CrateNettyHttpServerTransport;
+import io.crate.protocols.http.CrateNettyHttpServerTransport;
 import io.crate.operation.auth.Authentication;
 import io.crate.operation.auth.AuthenticationMethod;
 import io.crate.operation.auth.HbaProtocol;
 import io.crate.operation.user.User;
-import io.crate.protocols.postgres.ssl.SslReqHandler;
 import io.crate.protocols.postgres.types.PGType;
 import io.crate.protocols.postgres.types.PGTypes;
 import io.crate.types.DataType;
