+++ users/src/main/java/io/crate/operation/auth/UserServiceFactoryImpl.java
@@ -22,7 +22,7 @@
 
 package io.crate.operation.auth;
 
-import io.crate.http.netty.HttpAuthUpstreamHandler;
+import io.crate.protocols.http.HttpAuthUpstreamHandler;
 import io.crate.metadata.sys.SysUsersTableInfo;
 import io.crate.operation.collect.sources.SysTableRegistry;
 import io.crate.operation.user.*;
